pyproject.toml
[tool.poetry]
name = "async-task-api"
version = "0.1.0"
description = "Async Task Management API with FastAPI, async SQLAlchemy, PostgreSQL, Alembic"
authors = ["Rade Stojkovic <rade.stojkovic991@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
sqlalchemy = {version = "^2.0.0", extras = ["asyncio"]}
alembic = "^1.13.1"
pydantic = "^2.6.0"
pytest = "^8.0.0"
httpx = "^0.27.0"
psycopg2-binary = "^2.9.9" 
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
